refactor: isolate inputs as component
