+++ src/components/layouts/Router.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createBrowserRouter, Route, createRoutesFromElements } from 'react-router-dom';
 import Home from '../main/home/Home';
-import Forms from '../main/forms/Forms';
+import FormPage from '../main/forms/FormPage';
 import Title from '../header/Title';
 import RootLayout from './RootLayout';
 
@@ -10,7 +10,7 @@ const router = createBrowserRouter(
     <Route path="/" element={<RootLayout />}>
       <Route index element={<Home />}></Route>
       <Route path="about" element={<Title>{'About'}</Title>}></Route>
-      <Route path="forms" element={<Forms />}></Route>
+      <Route path="forms" element={<FormPage />}></Route>
       <Route path="*" element={<Title>{'Page not found :('}</Title>}></Route>
     </Route>
   )
