+++ src/components/main/forms/Form.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { fireEvent, render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import Input from './Input';
+import Input from './Form';
 
 const mockNames = {
   firstName: ['Jack', 'John'],
